Move the rest of metabot's endpoints into metabot/api.ts
api/metabot.ts is deleted; every metabot endpoint now lives in the
metabot module's api.ts as one injectEndpoints call, and all consumers
(metabot itself, admin/ai, EE, and the SDK) import from
metabase/metabot/api instead of the metabase/api barrel.
