+++ enterprise/frontend/src/embedding-sdk-ee/metabot/MetabotChatSuggestions.tsx
@@ -1,4 +1,4 @@
-import { useGetSuggestedMetabotPromptsQuery } from "metabase/api";
+import { useGetSuggestedMetabotPromptsQuery } from "metabase/metabot/api";
 import { useMetabotAgent } from "metabase/metabot/hooks";
 import { Button, Stack } from "metabase/ui";
