+++ enterprise/frontend/src/metabase-enterprise/metabot/components/MetabotAdmin/MetabaseAIProviderSetup.tsx
@@ -2,12 +2,10 @@ import { useCallback, useEffect, useState } from "react";
 import { match } from "ts-pattern";
 import { jt, t } from "ttag";
 
-import {
-  useRefreshTokenStatusMutation,
-  useUpdateMetabotSettingsMutation,
-} from "metabase/api";
+import { useRefreshTokenStatusMutation } from "metabase/api";
 import { getErrorMessage } from "metabase/api/utils";
 import { useAIProviderConfigurationContext } from "metabase/metabot";
+import { useUpdateMetabotSettingsMutation } from "metabase/metabot/api";
 import { MetabotManagedProviderLimitActions } from "metabase/metabot/components/MetabotManagedProviderLimit";
 import type { MetabaseAIProviderSetupProps } from "metabase/plugins";
 import { useSelector } from "metabase/redux";
