+++ enterprise/frontend/src/metabase-enterprise/moderation/components/MetabotVerifiedContentConfigurationPane.tsx
@@ -1,8 +1,8 @@
 import { t } from "ttag";
 
 import { SettingHeader } from "metabase/admin/settings/components/SettingHeader";
-import { useUpdateMetabotMutation } from "metabase/api";
 import { useToast } from "metabase/common/hooks";
+import { useUpdateMetabotMutation } from "metabase/metabot/api";
 import { Stack, Switch } from "metabase/ui";
 import type { MetabotInfo } from "metabase-types/api";
