+++ frontend/src/metabase/admin/ai/AIProviderSettingsSection.tsx
@@ -3,12 +3,13 @@ import { P, match } from "ts-pattern";
 import { t } from "ttag";
 
 import { SettingsSection } from "metabase/admin/components/SettingsSection";
-import { skipToken, useGetMetabotSettingsQuery } from "metabase/api";
+import { skipToken } from "metabase/api";
 import {
   AIProviderConfigurationForm,
   getProviderOptions,
   parseProviderAndModel,
 } from "metabase/metabot";
+import { useGetMetabotSettingsQuery } from "metabase/metabot/api";
 import { PLUGIN_METABOT } from "metabase/plugins";
 import { useAdminSetting, useSetting } from "metabase/settings";
 import { Badge, Flex, Group } from "metabase/ui";
