+++ frontend/src/metabase/admin/ai/MetabotSettingsPanel.tsx
@@ -4,16 +4,13 @@ import { match } from "ts-pattern";
 import { c, t } from "ttag";
 
 import { SettingHeader } from "metabase/admin/settings/components/SettingHeader";
-import {
-  skipToken,
-  useGetCollectionQuery,
-  useUpdateMetabotMutation,
-} from "metabase/api";
+import { skipToken, useGetCollectionQuery } from "metabase/api";
 import { canonicalCollectionId } from "metabase/common/collections/utils";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
 import { CollectionPickerModal } from "metabase/common/components/Pickers/CollectionPicker";
 import { useToast } from "metabase/common/hooks";
 import { useGetIcon } from "metabase/hooks/use-icon";
+import { useUpdateMetabotMutation } from "metabase/metabot/api";
 import { FIXED_METABOT_ENTITY_IDS } from "metabase/metabot/constants";
 import { PLUGIN_MODERATION } from "metabase/plugins";
 import { useAdminSetting } from "metabase/settings";
