+++ frontend/src/metabase/metabot/components/AIProviderConfigurationForm/AzureProviderFields.tsx
@@ -2,7 +2,6 @@ import { type FormikHelpers, useFormikContext } from "formik";
 import { useMemo } from "react";
 import { c, t } from "ttag";
 
-import { useUpdateMetabotSettingsMutation } from "metabase/api";
 import { ExternalLink } from "metabase/common/components/ExternalLink";
 import { SetByEnvVar } from "metabase/common/components/SetByEnvVar";
 import {
@@ -11,6 +10,7 @@ import {
   FormSelect,
   FormTextInput,
 } from "metabase/forms";
+import { useUpdateMetabotSettingsMutation } from "metabase/metabot/api";
 import { useAdminSettings } from "metabase/settings";
 
 import { useAIProviderConfigurationContext } from "./AIProviderConfigurationContext";
