+++ frontend/src/metabase/metabot/components/AIProviderConfigurationForm/BedrockProviderFields.tsx
@@ -2,10 +2,10 @@ import { type FormikHelpers, useFormikContext } from "formik";
 import { useMemo } from "react";
 import { t } from "ttag";
 
-import { useUpdateMetabotSettingsMutation } from "metabase/api";
 import { ExternalLink } from "metabase/common/components/ExternalLink";
 import { SetByEnvVar } from "metabase/common/components/SetByEnvVar";
 import { FormErrorMessage, FormProvider, FormTextInput } from "metabase/forms";
+import { useUpdateMetabotSettingsMutation } from "metabase/metabot/api";
 import { useAdminSettings } from "metabase/settings";
 import { Text } from "metabase/ui";
 import type { BedrockCredentials } from "metabase-types/api";
