+++ frontend/src/metabase/metabot/components/Metabot.tsx
@@ -4,8 +4,8 @@ import { t } from "ttag";
 
 import MetabotFailure from "assets/img/metabot-failure.svg?component";
 import ErrorBoundary from "metabase/ErrorBoundary";
-import { metabotApi } from "metabase/api";
 import { idTag } from "metabase/api/tags";
+import { metabotApi } from "metabase/metabot/api";
 import {
   useIsFullPageMetabot,
   useMetabotAgent,
