+++ frontend/src/metabase/metabot/components/MetabotChat/MetabotAgentDataSourcePills.tsx
@@ -8,12 +8,12 @@ import {
   useGetDatabaseQuery,
   useGetFieldTableIdsQuery,
   useGetTableQuery,
-  useSubmitMetabotSourceFeedbackMutation,
 } from "metabase/api";
 import type { GeneratedCard } from "metabase/api/ai-streaming/schemas";
 import { ForwardRefLink } from "metabase/common/components/Link";
 import { useToast } from "metabase/common/hooks";
 import { deserializeCardFromQuery } from "metabase/common/utils/card";
+import { useSubmitMetabotSourceFeedbackMutation } from "metabase/metabot/api";
 import { getMetabotId } from "metabase/metabot/state";
 import {
   getCollectionLocationLabel,
