+++ frontend/src/metabase/metabot/components/MetabotChat/MetabotChatMessage.tsx
@@ -5,8 +5,8 @@ import { Fragment, forwardRef, useCallback, useMemo, useState } from "react";
 import { match } from "ts-pattern";
 import { t } from "ttag";
 
-import { useSubmitMetabotFeedbackMutation } from "metabase/api/metabot";
 import { useToast } from "metabase/common/hooks";
+import { useSubmitMetabotFeedbackMutation } from "metabase/metabot/api";
 import { MetabotManagedProviderLimitActions } from "metabase/metabot/components/MetabotManagedProviderLimit";
 import {
   type MetabotAgentChatMessage,
