+++ frontend/src/metabase/metabot/components/MetabotChat/MetabotInlineChart.tsx
@@ -10,11 +10,11 @@ import {
   useGetCardQuery,
 } from "metabase/api";
 import type { GeneratedCard } from "metabase/api/ai-streaming/schemas";
-import { useSaveMetabotEntityMutation } from "metabase/api/metabot";
 import { ForwardRefLink } from "metabase/common/components/Link";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
 import { SaveQuestionModal } from "metabase/common/components/SaveQuestionModal";
 import { serializeChartClipboard } from "metabase/common/utils/chart-clipboard";
+import { useSaveMetabotEntityMutation } from "metabase/metabot/api";
 import { getSavedChartCardId, markChartSaved } from "metabase/metabot/state";
 import { useDispatch, useSelector } from "metabase/redux";
 import { addUndo } from "metabase/redux/undo";
